A short's becoming a little loopy

+++ powercraft/api/utils/MathHelper.java
@@ -20,7 +20,7 @@ public class MathHelper {
 	    long i = Double.doubleToLongBits(x);
 	    i = 0x5fe6ec85e7de30daL - (i>>1);
 	    x = Double.longBitsToDouble(i);
-	    for(int it = 0; it < 4; it++){
+	    for(short it = 0; it < 4; it++){
 	    	x = x*(1.5d - xhalf*x*x);
 	    }
 	    x *= number;
